Refactor: Extract useRouletteSettings custom hook to centralize state management. This change creates a new custom hook `useRouletteSettings` to handle the state logic for the roulette's title and items. The hook encapsulates `useState` and the `addItem`, `removeItem`, and `updateItem` functions. The following three components have been refactored to use this new hook, removing duplicated code and improving maintainability: `original-roulette/Client.tsx`, `mypage/roulette/[id]/Client.tsx`, and `templates/roulette/[id]/Client.tsx`.
